[Added] Unit tests over UTCalculationMethods. [Added] Integration tests over UTCalculationFactory. [Refactored] Calculation method MultiplicationPerAddition

+++ UICalculationResultSubscriber.pas
@@ -5,7 +5,7 @@ interface
   type
     ICalculationResultSubscriber = interface
     ['{BC2D58A7-6316-45ED-99C3-F3E7A420EB58}']
-      procedure Update(aCalculationResult: Integer);
+      procedure Update(aCalculationResult: Int64);
     end;
 
 implementation
